% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Distance and damage of fire.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{distance}{distance the fire spread in miles}
  \item{damage}{damage done in millions}
  ...
}
}
\source{
\url{"https://www.google.com/"}
}
\usage{
fire
}
\description{
A dataset containing the distance and damage of a fire.
}
\keyword{datasets}
