% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{histogram}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iteration}

\item{x}{sample}

\item{fun}{which function to sample}

\item{alpha}{confidence interval = 1-a}

\item{cx}{scale of graph}

\item{...}{extra arguments for plot}
}
\value{
a list of info about the plot and a histogram
}
\description{
histogram
}
\examples{
sam = c(10, 20, 5, 10, 2, 5, 15)
myboot2(10000, sam, "mean", 0.05)

}
