% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Mean Confidence Interval}
\usage{
myci(x)
}
\arguments{
\item{x}{Vector of values}
}
\value{
Confidence Interval
}
\description{
Mean Confidence Interval
}
\examples{
x = c(5, 4, 3, 5, 7, 1)
myci(x)

}
