% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poisson Distibution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{iter}

\item{lambda}{probability}

\item{...}{}
}
\value{
a plot of the poisson distibution
}
\description{
Poisson Distibution
}
\examples{
mycltp(2, 10000, 4)

}
