% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{normal max likelihood estimator}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{vector of data}

\item{mu}{mean of population}

\item{sig}{standard hdviation of population}

\item{...}{variables for the plot}
}
\value{
list of values for max likelihood
}
\description{
normal max likelihood estimator
}
\examples{
x=1:10
mymlnorm(x=x,mu=seq(10,15,length=1000),sig=seq(0.1,4,length=1000))

}
