% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal curve plot}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean value}

\item{sigma}{standard variation}

\item{a}{probability constant}
}
\value{
normal graph and probability
}
\description{
Creates a plot of the normal distrubuition and shows the probability as the area.
}
\examples{
 myncurve(10, 5, 6)

}
