% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{p value plot}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{tcalc}

\item{xmax}{max x value on plot}

\item{n}{number of iterations}

\item{alpha}{used for the confidence value}
}
\value{
list of values from plot & the plot
}
\description{
p value plot
}
\examples{
tcalc = 1.2
mypvalue(tcalc)

}
