% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{T Test}
\usage{
myttest(x, conf.level = 0.95)
}
\arguments{
\item{x}{vector of values}

\item{conf.level}{confidence level}
}
\value{
t value
}
\description{
T Test
}
\examples{
x=1:30
myttest(x)

}
